clean up and bug fix of utilization constraints and adding jake's dnvgl c202 code
