+++ examples/05_tower_monopile/monopile_direct.py
@@ -40,6 +40,7 @@ modeling_options["materials"] = {}
 modeling_options["WISDEM"] = {}
 modeling_options["WISDEM"]["TowerSE"] = {}
 modeling_options["WISDEM"]["TowerSE"]["buckling_length"] = 30.0
+modeling_options["WISDEM"]["TowerSE"]["buckling_method"] = "euro"
 modeling_options["flags"]["monopile"] = True
 
 # Monopile foundation
