+++ examples/05_tower_monopile/tower_direct.py
@@ -30,6 +30,7 @@ modeling_options["flags"] = {}
 modeling_options["materials"] = {}
 modeling_options["WISDEM"] = {}
 modeling_options["WISDEM"]["TowerSE"] = {}
+modeling_options["WISDEM"]["TowerSE"]["buckling_method"] = "dnvgl"
 modeling_options["WISDEM"]["TowerSE"]["buckling_length"] = 30.0
 modeling_options["flags"]["monopile"] = False
